ecluindo arquivos obsoletos
