+++ README.md
@@ -1,2 +1,30 @@
 # first-docker-MERN
 This is my first MERN project, it is made in docker environment with React JS, Mongodb, Node JS. Is a very simple movie list app.
+
+## Para rodar este projeto após cloná-lo
+
+### No diretório backend importar
+*express
+*mongoose
+*nodemon
+*body-parser 
+*cors
+
+### No diretório frontend importar
+*Axios
+*Bootstrap
+*StyledComponents
+*react-table-6 (Isto é importante, a versão deve ser a 6)
+*react-router-dom
+
+## Após as importações
+
+### No diretório raíz rodar no terminal
+
+*docker compose build
+*docker compose up -d
+
+### no diretório frontend rodar no terminal
+
+yarn start ou npm start
+
