--- backend/db/index.js
@@ -1,14 +1,0 @@
-/*
-const mongoose = require('mongoose')
-
-mongoose
-    .connect('mongodb://127.0.0.1:27017/cinema', { useNewUrlParser: true })
-    .catch(e => {
-        console.error('Connection error', e.message)
-    })
-
-const db = mongoose.connection
-
-module.exports = db
-
-*/
